lib/csp,main,adcs: Move update API for CSP stat
This commit moves the API that updates the CSP received counter, the
last received command ID, and related functions to the library.
